+++ include/ctk_string.h
@@ -84,7 +84,7 @@ int ctk_strerror_r(int errnum, char * result, size_t length);
     ctk_strpos()
 ------------------------------------------------------------------------------*/
 /**
- * Finds the first occurrence of \p{substr} in \p{str} and returns its position.
+ * Finds the first position of \p{substr} in \p{str}.
  *
  * - Terminating null characters are excluded from the search.
  * - Both \p{str} and \p{substr} can be string litterals.
@@ -103,6 +103,29 @@ int ctk_strerror_r(int errnum, char * result, size_t length);
  * - \b [UB] \p{substr} must be a pointer to a null-terminated string.
  */
 ptrdiff_t ctk_strpos(const char * str, const char * substr);
+/*------------------------------------------------------------------------------
+    ctk_strrpos()
+------------------------------------------------------------------------------*/
+/**
+ * Finds the last position of \p{substr} in \p{str}.
+ *
+ * - Terminating null characters are excluded from the search.
+ * - Both \p{str} and \p{substr} can be string litterals.
+ *
+ * @param[in] str    : The string to search in.
+ * @param[in] substr : The string to search for.
+ *
+ * @return
+ * - @success:
+ *   - The position in \p{srt} of the first character from \p{substr}.
+ *   - '0' if \p{substr} is an empty string.
+ * - @failure: `-1`.
+ *
+ * @warning
+ * - \b [UB] \p{str} must be a pointer to a null-terminated string.
+ * - \b [UB] \p{substr} must be a pointer to a null-terminated string.
+ */
+ptrdiff_t ctk_strrpos(const char * str, const char * substr);
 /*------------------------------------------------------------------------------
     ctk_strrstr()
 ------------------------------------------------------------------------------*/